+++ NINA.Image/FileFormat/FITS/CfitsioNative.cs
@@ -102,7 +102,7 @@ namespace NINA.Image.FileFormat.FITS {
         }
 
         public static void LogErrorStatus(string op, int statusCode) {
-            if(statusCode != 0) {
+            if (statusCode != 0) {
                 Logger.Error($"{op} failed with {statusCode} = {fits_get_errstatus(statusCode)}");
             }
         }
@@ -155,7 +155,7 @@ namespace NINA.Image.FileFormat.FITS {
         public static extern int fits_read_pix(
             IntPtr fptr,
             DATATYPE datatype,
-            int[] firstpix,
+            long[] firstpix,
             long nelem,
             IntPtr nulval,
             IntPtr array,
@@ -379,7 +379,7 @@ namespace NINA.Image.FileFormat.FITS {
         }
 
         public static T[] read_pixels<T>(IntPtr fptr, int naxes, int nelem) where T : unmanaged {
-            var firstpix = new int[naxes];
+            var firstpix = new long[naxes];
             for (int i = 0; i < naxes; i++) {
                 firstpix[i] = 1;
             }
